move kebab-menu-actions
