+++ src/modules/feature/board-collaborative-text-editor-element/components/CollaborativeTextEditorElementMenu.unit.ts
@@ -7,7 +7,7 @@ import {
 	KebabMenuActionDelete,
 	KebabMenuActionMoveDown,
 	KebabMenuActionMoveUp,
-} from "@ui-board";
+} from "@ui-kebab-menu";
 import { useDeleteConfirmationDialog } from "@ui-confirmation-dialog";
 import { shallowMount } from "@vue/test-utils";
 import { ref } from "vue";
