+++ src/modules/feature/media-shelf/MediaBoardExternalToolDeletedElement.unit.ts
@@ -3,7 +3,7 @@ import {
 	createTestingI18n,
 	createTestingVuetify,
 } from "@@/tests/test-utils/setup";
-import { KebabMenuAction } from "@ui-board";
+import { KebabMenuAction } from "@ui-kebab-menu";
 import { mount } from "@vue/test-utils";
 import { nextTick } from "vue";
 import { ComponentProps } from "vue-component-type-helpers";
