+++ src/modules/ui/board/index.ts
@@ -1,34 +1,11 @@
 import BoardMenu from "./BoardMenu.vue";
-import KebabMenuAction from "./KebabMenuAction.vue";
-import KebabMenuActionCopy from "./KebabMenuActionCopy.vue";
-import KebabMenuActionDelete from "./KebabMenuActionDelete.vue";
-import KebabMenuActionEdit from "./KebabMenuActionEdit.vue";
-import KebabMenuActionMoveDown from "./KebabMenuActionMoveDown.vue";
-import KebabMenuActionMoveLeft from "./KebabMenuActionMoveLeft.vue";
-import KebabMenuActionMoveRight from "./KebabMenuActionMoveRight.vue";
-import KebabMenuActionMoveUp from "./KebabMenuActionMoveUp.vue";
-import KebabMenuActionPublish from "./KebabMenuActionPublish.vue";
-import KebabMenuActionRevert from "./KebabMenuActionRevert.vue";
-import KebabMenuActionShare from "./KebabMenuActionShare.vue";
-import KebabMenuActionRename from "./KebabMenuActionRename.vue";
+import { BoardMenuScope } from "./board-menu-scope";
 import BoardSectionCreationHeader from "./BoardSectionCreationHeader.vue";
 import ContentElementBar from "./content-element/ContentElementBar.vue";
 
-export { BoardMenuScope } from "./board-menu-scope";
 export {
 	BoardMenu,
-	KebabMenuAction,
-	KebabMenuActionCopy,
-	KebabMenuActionEdit,
-	KebabMenuActionDelete,
-	KebabMenuActionMoveUp,
-	KebabMenuActionMoveDown,
-	KebabMenuActionMoveLeft,
-	KebabMenuActionMoveRight,
-	KebabMenuActionPublish,
-	KebabMenuActionRevert,
-	KebabMenuActionShare,
-	KebabMenuActionRename,
+	BoardMenuScope,
 	ContentElementBar,
 	BoardSectionCreationHeader,
 };
