+++ src/modules/ui/kebab-menu/KebabMenuActionEdit.vue
@@ -8,8 +8,8 @@
 </template>
 
 <script setup lang="ts">
+import KebabMenuAction from "./KebabMenuAction.vue";
 import { mdiPencilOutline } from "@icons/material";
-import { KebabMenuAction } from "@ui-board";
 import { useI18n } from "vue-i18n";
 const { t } = useI18n();
 </script>
