+++ src/modules/ui/kebab-menu/KebabMenuActionRevert.vue
@@ -5,8 +5,8 @@
 </template>
 
 <script setup lang="ts">
+import KebabMenuAction from "./KebabMenuAction.vue";
 import { mdiRestore } from "@icons/material";
-import { KebabMenuAction } from "@ui-board";
 import { useI18n } from "vue-i18n";
 const { t } = useI18n();
 </script>
