+++ src/modules/ui/kebab-menu/index.ts
@@ -1,3 +1,29 @@
 import KebabMenu from "./KebabMenu.vue";
+import KebabMenuAction from "./KebabMenuAction.vue";
+import KebabMenuActionCopy from "./KebabMenuActionCopy.vue";
+import KebabMenuActionDelete from "./KebabMenuActionDelete.vue";
+import KebabMenuActionEdit from "./KebabMenuActionEdit.vue";
+import KebabMenuActionMoveDown from "./KebabMenuActionMoveDown.vue";
+import KebabMenuActionMoveLeft from "./KebabMenuActionMoveLeft.vue";
+import KebabMenuActionMoveRight from "./KebabMenuActionMoveRight.vue";
+import KebabMenuActionMoveUp from "./KebabMenuActionMoveUp.vue";
+import KebabMenuActionPublish from "./KebabMenuActionPublish.vue";
+import KebabMenuActionRevert from "./KebabMenuActionRevert.vue";
+import KebabMenuActionShare from "./KebabMenuActionShare.vue";
+import KebabMenuActionRename from "./KebabMenuActionRename.vue";
 
-export { KebabMenu };
+export {
+	KebabMenu,
+	KebabMenuAction,
+	KebabMenuActionCopy,
+	KebabMenuActionEdit,
+	KebabMenuActionDelete,
+	KebabMenuActionMoveUp,
+	KebabMenuActionMoveDown,
+	KebabMenuActionMoveLeft,
+	KebabMenuActionMoveRight,
+	KebabMenuActionPublish,
+	KebabMenuActionRevert,
+	KebabMenuActionShare,
+	KebabMenuActionRename,
+};
